Refactor: sort out lensPragmaOptions by new module Interaction.Options.Types
Problem addressed:
```haskell
-- | Defined locally here since module ''Agda.Interaction.Options.Lenses''
--   has cyclic dependency.
lensPragmaOptions :: Lens' CommandLineOptions PragmaOptions
```

The solution is to extract option types and LensPragmaOptions into a
new boot module Agda.Interaction.Options.Types.
